Changes in the services , add stages to pipleine, and changes in controllers

+++ src/controllers/area.controller.ts
@@ -4,6 +4,7 @@ import { TYPES } from "../constants/TYPES";
 import { AreaService } from "../services/area.service";
 import { Request, Response } from "express";
 import { IArea } from "../interfaces/IArea";
+import { isValidObjectId } from "mongoose";
 
 @controller("/area")
 export class AreaController {
@@ -14,21 +15,20 @@ export class AreaController {
         try {
             const { areaName } = req.body;
             await this.areaService.addArea({ areaName })
-            res.status(201).json({ message: "Area Added Successfully" });
+            res.status(201).json({ status: true, message: "Area Added Successfully" });
         } catch (error) {
-            console.log(error);
-            res.status(500).json({ message: error.message });
+            res.status(500).json({ status: false, message: error.message });
         }
     }
 
     @httpGet("/getAllAreas")
     async getAllModules(req: Request, res: Response) {
         try {
-            const {page,limit}=req.query;
-            const data = await this.areaService.getAllAreas({page,limit});
-            return res.status(201).json({ data: data })
+            const { page, limit } = req.query;
+            const data = await this.areaService.getAllAreas({ page, limit });
+            return res.status(201).json({ status: true, data: data })
         } catch (error) {
-            return res.status(500).json({ message: error.message });
+            return res.status(500).json({ status: false, message: error.message });
         }
     }
 
@@ -36,21 +36,30 @@ export class AreaController {
     async changeSequence(req: Request, res: Response) {
         try {
             const { areaId, sequence } = req.body;
+            if (!isValidObjectId(areaId)) {
+                throw new Error("Not Valid Area Id")
+            }
+            if (!sequence || sequence <=0 ) {
+                throw new Error("please enter Sequence")
+            }
             await this.areaService.changeSequence({ areaId, sequence } as IArea);
-            return res.status(201).json({ message: "Sequence Changed Succesfully" })
+            return res.status(201).json({ status: true, message: "Sequence Changed Succesfully" })
         } catch (error) {
-            return res.status(500).json({ message: error.message });
+            return res.status(500).json({ status: false, message: error.message });
         }
     }
 
-    @httpDelete("/deleteArea")
+    @httpDelete("/deleteArea/:id")
     async deleteHomePageData(req: Request, res: Response) {
         try {
-            const { areaId } = req.body;
+            const areaId = req.params.id;
+            if (!isValidObjectId(areaId)) {
+                throw new Error("Not Valid Area Id")
+            }
             await this.areaService.deleteArea(areaId);
-            return res.status(201).json({ message: "Data Deleted Succesfully" })
+            return res.status(201).json({ status: true, message: "Data Deleted Succesfully" })
         } catch (error) {
-            return res.status(500).json({ message: error.message });
+            return res.status(500).json({ status: false, message: error.message });
         }
     }
 }
